data-client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-client"
version = "0.1.0"
description = "grpc client for streaming data to kafka producer server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio-tools>=1.71.0",
    "protoletariat>=3.3.10",
    "protobuf>=4.24.4",
    "kafka-python>=2.0.2",
    "python-dotenv>=1.0.0",
    "python-snappy>=0.6.1",
    "retrying>=1.3.4",
]
